+++ src/pages/CartList/Containers/ItemManage/styles.js
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled/macro';
+import { COLORS } from 'styles/theme';
+
+const ItemManageContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+
+  font-size: 0.875rem;
+  color: ${COLORS.GRAY_30};
+`;
+
+export { ItemManageContainer };
